#17627 #38336 Move util method from DBUtils to DBWUtils, add java doc comment
