+++ plugins/org.jkiss.dbeaver.model/src/org/jkiss/dbeaver/model/net/DBWUtils.java
@@ -18,9 +18,17 @@ package org.jkiss.dbeaver.model.net;
 
 import org.jkiss.code.NotNull;
 import org.jkiss.code.Nullable;
+import org.jkiss.dbeaver.DBException;
+import org.jkiss.dbeaver.model.DBConstants;
 import org.jkiss.dbeaver.model.DBPDataSourceContainer;
+import org.jkiss.dbeaver.model.DatabaseURL;
 import org.jkiss.dbeaver.model.connection.DBPConnectionConfiguration;
+import org.jkiss.dbeaver.model.connection.DBPDriver;
 import org.jkiss.utils.CommonUtils;
+
+import java.net.URI;
+import java.util.Collections;
+import java.util.Set;
 
 public class DBWUtils {
 
@@ -101,6 +109,106 @@ public class DBWUtils {
         return CommonUtils.isEmpty(cfg.getConfigProfileName())
             ? null
             : dataSourceContainer.getRegistry().getNetworkProfile(cfg.getConfigProfileSource(), cfg.getConfigProfileName());
+    }
+
+
+    public record ConnectivityParameters(
+        @Nullable String hostName,
+        @Nullable String hostPort,
+        @Nullable String databaseName,
+        @Nullable String userName,
+        @Nullable String server
+    ) {
+    }
+
+    @NotNull
+    private static ConnectivityParameters getExplicitConnectivityParameters(@NotNull DBPConnectionConfiguration configuration) {
+        String defaultCatalogName = configuration.getBootstrap().getDefaultCatalogName();
+        return new ConnectivityParameters(
+            configuration.getHostName(),
+            configuration.getHostPort(),
+            CommonUtils.isNotEmpty(defaultCatalogName) ? defaultCatalogName : configuration.getDatabaseName(),
+            configuration.getUserName(),
+            configuration.getServerName()
+        );
+    }
+
+    /**
+     * Returns information about connection by its configuration.
+     * If the configuration type is URL, it extracts information
+     * according to the sample URL template in the driver properties
+     * or generic URL template, if sample URL template is empty.
+     */
+    @NotNull
+    public static ConnectivityParameters getConnectivityParameters(
+        @NotNull DBPConnectionConfiguration configuration,
+        @NotNull DBPDriver driver
+    ) throws DBException {
+        ConnectivityParameters explicitConfiguration = getExplicitConnectivityParameters(configuration);
+        return switch (configuration.getConfigurationType()) {
+            case MANUAL -> explicitConfiguration;
+            case URL -> {
+                String activeUrl = driver.getConnectionURL(configuration);
+                if (CommonUtils.isNotEmpty(activeUrl)) {
+                    ConnectivityParameters urlConnectivityParams = null;
+                    DBPConnectionConfiguration urlConfiguration = null;
+                    DatabaseURL.MetaURL metaURL = null;
+                    if (CommonUtils.isNotEmpty(driver.getSampleURL())) {
+                        urlConfiguration = DatabaseURL.extractConfigurationFromUrl(driver.getSampleURL(), activeUrl);
+                        if (urlConfiguration != null) {
+                            metaURL = DatabaseURL.parseSampleURL(driver.getSampleURL());
+                        }
+                    }
+                    if (urlConfiguration == null) {
+                        urlConfiguration = DatabaseURL.extractConfigurationFromUrl(DatabaseURL.GENERIC_URL_TEMPLATE, activeUrl);
+                        if (urlConfiguration != null) {
+                            metaURL = DatabaseURL.parseSampleURL(DatabaseURL.GENERIC_URL_TEMPLATE);
+                        }
+                    }
+                    if (urlConfiguration != null) {
+                        urlConnectivityParams = getExplicitConnectivityParameters(urlConfiguration);
+                    }
+                    if (urlConnectivityParams == null) {
+                        final String jdbcPrefix = "jdbc:";
+                        URI url = URI.create(activeUrl.startsWith(jdbcPrefix) ? activeUrl.substring(jdbcPrefix.length()) : activeUrl);
+                        urlConnectivityParams = new ConnectivityParameters(
+                            url.getHost(),
+                            url.getPort() != -1 ? Integer.toString(url.getPort()) : null,
+                            url.getPath().startsWith("/") ? url.getPath().substring(1) : url.getPath(),
+                            url.getUserInfo(),
+                            null
+                        );
+                    }
+                    Set<String> requiredUrlParts = metaURL != null ? metaURL.getRequiredProperties() : Collections.emptySet();
+
+                    String databaseName = requiredUrlParts.contains(DBConstants.PROP_DATABASE)
+                        ? urlConnectivityParams.databaseName()
+                        : CommonUtils.isNotEmpty(urlConnectivityParams.databaseName())
+                            ? urlConnectivityParams.databaseName()
+                            : explicitConfiguration.databaseName();
+                    String userName = requiredUrlParts.contains(DBConstants.PROP_USER)
+                        ? urlConnectivityParams.userName()
+                        : CommonUtils.isNotEmpty(urlConnectivityParams.userName())
+                            ? urlConnectivityParams.userName()
+                            : explicitConfiguration.userName();
+                    yield new ConnectivityParameters(
+                        urlConnectivityParams.hostName(),
+                        urlConnectivityParams.hostPort(),
+                        databaseName,
+                        userName,
+                        urlConnectivityParams.server()
+                    );
+                } else {
+                    yield new ConnectivityParameters(
+                        driver.getDefaultHost(),
+                        driver.getDefaultPort(),
+                        explicitConfiguration.databaseName(),
+                        explicitConfiguration.userName(),
+                        null
+                    );
+                }
+            }
+        };
     }
 }
